Don't assume int is 32 bit

+++ MountainCheckHours.py
@@ -11,7 +11,7 @@ def find_savedir():
     winpath = path.expandvars(r"%USERPROFILE%\\AppData\\LocalLow\\David OReilly\\Mountain")
     if path.isdir(winpath):
         return winpath
-    
+
     linpath = path.expandvars(r"~/.steam/steam/steamapps/compatdata/897330/pfx/dosdevices/z:/home")
     linpath_x = "/.config/unity3d/David OReilly/Mountain"
     if path.isdir(linpath):
@@ -27,8 +27,8 @@ if __name__ == "__main__":
         print("Cannot find savegame. Please move this exe/script to your save folder for Mountain.")
     else:
         filename = "%s/%s" % (path, savefile)
-        a = numpy.fromfile(filename, dtype=int, count=-1, sep='')
+        a = numpy.fromfile(filename, dtype=numpy.int32, count=-1, sep='')
         hours_done = a[11]
-        print("%s of 987 hours complete" % hours_done) 
+        print("%s of 987 hours complete" % hours_done)
     print("")
     input("Press enter to exit program...")
